# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        maxLength: 255
        type: string
      password:
        example: password123
        maxLength: 128
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.StandardErrorResponse:
    properties:
      error:
        example: Invalid email or password
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.UpdateProfileRequest:
    properties:
      name:
        example: John Doe
        maxLength: 100
        minLength: 2
        type: string
      phone:
        example: "+1234567890"
        maxLength: 20
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
  /analytics/dashboard:
    get:
      consumes:
      - application/json
      description: Get comprehensive dashboard analytics including user statistics
        and recent interactions
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard analytics retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard analytics
      tags:
      - Analytics
  /analytics/export:
    get:
      consumes:
      - application/json
      description: Export user analytics data in JSON or CSV format
      parameters:
      - default: '"json"'
        description: Export format (json, csv)
        in: query
        name: format
        type: string
      - default: 30
        description: Number of days to export
        in: query
        name: days
        type: integer
      produces:
      - application/json
      - text/csv
      responses:
        "200":
          description: Analytics data exported successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Unsupported format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export analytics data
      tags:
      - Analytics
  /analytics/products:
    get:
      consumes:
      - application/json
      description: Get product performance analytics including top selling products,
        most viewed, and category performance
      parameters:
      - default: 30
        description: Number of days to analyze
        in: query
        name: days
        type: integer
      - default: 20
        description: Number of items to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product analytics retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get product analytics
      tags:
      - Analytics
...
  /analytics/search:
    get:
      consumes:
      - application/json
      description: Get search analytics and query performance metrics
      parameters:
      - default: 30
        description: Number of days to analyze
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search analytics retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get search analytics
      tags:
      - Analytics
  /analytics/trends:
    get:
      consumes:
      - application/json
      description: Get ML-powered trend analysis and predictions
      parameters:
      - default: 30
        description: Number of days to analyze
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ML trends retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get ML trends
      tags:
      - Analytics
  /analytics/user:
    get:
      consumes:
      - application/json
      description: Get detailed user analytics including interaction stats, category
        preferences, and spending patterns
      parameters:
      - default: 30
        description: Number of days to analyze
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User analytics retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user analytics
      tags:
      - Analytics
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      summary: User login
      tags:
      - Authentication
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's comprehensive profile information
        including statistics and recent activity
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            $ref: '#/definitions/handlers.UserProfileResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Authentication
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, name, and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "409":
          description: User with this email already exists
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /cart:
    get:
      consumes:
      - application/json
      description: Get the current user's shopping cart with all items and total
      produces:
      - application/json
      responses:
        "200":
          description: Cart retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get shopping cart
      tags:
      - Cart
  /cart/add:
    post:
      consumes:
      - application/json
      description: Add a product to the user's shopping cart with specified quantity
      parameters:
      - description: Item to add to cart
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item added to cart successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or insufficient stock
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/clear:
    delete:
      consumes:
      - application/json
      description: Remove all items from the user's shopping cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart cleared successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clear shopping cart
      tags:
      - Cart
  /cart/item/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a specific item from the user's shopping cart
      parameters:
      - description: Cart Item ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item removed from cart successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid item ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart item not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update the quantity of a specific item in the user's cart
      parameters:
      - description: Cart Item ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New quantity (0 to remove item)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cart item updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or insufficient stock
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart item not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
  /comments:
    post:
      consumes:
      - application/json
      description: Add a comment and rating to a product
      parameters:
      - description: Comment to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment added successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add product comment
      tags:
      - Comments
  /comments/{comment_id}:
    delete:
      consumes:
      - application/json
      description: Delete a user's own comment
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid comment ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a user's own comment
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: comment_id
        required: true
        type: string
      - description: Updated comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized to update this comment
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - Comments
  /comments/{product_id}:
    get:
      consumes:
      - application/json
      description: Get paginated list of comments for a product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID
          schema:
            additionalProperties: true
            type: object
      summary: Get product comments
      tags:
      - Comments
  /discounts:
    post:
      consumes:
      - application/json
      description: Create a new product or category discount (admin only)
      parameters:
      - description: Discount data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDiscountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Discount created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create discount
      tags:
      - Discounts
  /discounts/active:
    get:
      consumes:
      - application/json
      description: Get list of currently active discounts
      parameters:
      - description: Filter by product ID
        in: query
        name: product_id
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Active discounts retrieved successfully
          schema:
            additionalProperties: true
            type: object
      summary: Get active discounts
      tags:
      - Discounts
...
  /orders/stats:
    get:
      consumes:
      - application/json
      description: Get comprehensive order statistics for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Order statistics retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order statistics
      tags:
      - Orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...